Reorganizing code/Added ability to add chance of genes to PawnKindDefs
