+++ Source/AllModdingComponents/JecsTools/BuildingExtension/PlaceWorker_Outline.cs
