+++ Source/AllModdingComponents/JecsTools/DamageSoak/HediffCompProperties_DamageSoak.cs
