+++ Source/AllModdingComponents/JecsTools/DamageSoak/HediffComp_DamageSoak.cs
