+++ Source/AllModdingComponents/JecsTools/ExtraMeleeDamages/HediffCompProperties_DamageOverTime.cs
