+++ Source/AllModdingComponents/JecsTools/ExtraMeleeDamages/HediffCompProperties_ExtraMeleeDamages.cs
