+++ Source/AllModdingComponents/JecsTools/ExtraMeleeDamages/Hediff_InjuryCustomLabel.cs
