+++ Source/AllModdingComponents/JecsTools/StartWithHediff/HediffExpandedDef.cs
