+++ Source/AllModdingComponents/JecsTools/StartWithHediff/HediffWithComps_Expanded.cs
